Moved classes out of Extensions namespace, and matched the namespaces for the types they're extending
